parallelized qsort
the quick sort is now parallelized. when the depth is less than 2 when it divides the vector in two it also spawns a thread to solve the other path (can create up to 4 threads). all of the threads run in parallel to solve their parts of the vector. the results are merged together using reduction for efficiency.