fixed function name
i was mistakenly making a merge sort algorithm so i renamed it as such. in addition i gave more control over the number of threads. the function asks for how many additional threads you want it to spawn and it will spawn them as evenly as possible to make sort as fast as possible.