dont think this will work
the goal of this branch is to use vector slices instead of making multiple copies of the vector but it seems that rust cant do this with the borrow checker. might come back to this later to try more stuff but for now this doesnt seem possible from what i know.